Made MultiPoint serializable in hdf5
